+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changes in sphinx-automodapi
 0.17.0 (unreleased)
 -------------------
 
+- Fixes issue where ``__slots__`` hides class variables. [#181]
+
 - Minimum supported Python version is now 3.8. [#177]
 
 0.16.0 (2023-08-17)